backend/backend/requirements.txt
aiohttp==3.7.4.post0
argon2-cffi==20.1.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
bcrypt==3.2.0
boto3==1.17.109
botocore==1.20.109
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
dj-database-url==0.5.0
-e git+http://github.com/anushka567/django.git@6a5ef557f80a8eb6a758ebe99c8bb477ca47459e#egg=Django
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-storages==1.11.1
docutils==0.17.1
geoip2==4.2.0
gunicorn==20.1.0
idna==2.10
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
maxminddb==2.0.3
multidict==5.1.0
numpy==1.21.0
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
pymemcache==3.4.4
python-dateutil==2.8.1
python-memcached==1.59
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.4.2
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
tblib==1.7.0
typing-extensions==3.10.0.0
tzdata==2021.1
urllib3==1.26.6
whitenoise==5.2.0
yarl==1.6.3
